Utilities moved to util
About:
- logSerial, setupBuzzer, buzz and waitForPiSerial go to util
- delete config.cpp
- change Serial.print to logSerial
- deactivate printing while flying
- rename BZZZ_VERBOSITY to BZZZ_LOGGING_LEVEL

Issues:
- Addresses #41
