+++ platformio.ini
@@ -24,16 +24,16 @@ board_build.f_cpu = 240000000L
 ; Build flags
 ; Set any preprocessor directives here
 ;
-; Use BZZZ_VERBOSITY to choose the level of verbosity
+; Use BZZZ_LOGGING_LEVEL to choose the level of verbosity
 ;  1: Debug
 ;  2: Info
 ;  3: Severe (only extremely critical information will be printed)
-; If BZZZ_VERBOSITY is not defined, the default is 3.
+; If BZZZ_LOGGING_LEVEL is not defined, the default is 3.
 ;
 ; BZZZ_DEBUG should be used only for debugging
 build_flags =
 	-D BZZZ_DEBUG
-	-D BZZZ_VERBOSITY=2
+	-D BZZZ_LOGGING_LEVEL=3
 
 ; Dependencies
 lib_deps =
